Replaced the fake visualiser with the pause icon.
I was always annoyed by this dancing little sh.t, whose only purpose is
to be like "Look at me! I'm moving for no good reason!".
Does it represent frequencies? No. Does it represent rhythm? No. Does it
distract you from staring at the playing track title while enjoying
music? Yes.
I suck at css and this took me a lot of trial and error, so it might not
be the best way to do it. But it the diff is small, and it works for me.